requirements.txt
fastapi
uvicorn
gunicorn
pydantic[email]
psycopg2
sqlalchemy
passlib
boto3
python-slugify
python-dotenv